Removed Format::Raw, because Format::Bash is really what I want
Json and TOML are still there. I need to do more serious work on
supporting json, tbh.

Improved the handling of not-found items and documented the behavior
in the readme. Expanded the examples in the readme.

Added formatting for boolean output: it's 0 and 1 for bash output.
Lastly, renamed OutputMode to Format, because that's what I kept calling it.