Add the ability to set booleans and numbers
This is the default for values that look like bools and numbers
on the command line. To coerce them to strings, wrap two layers
of quotes around them. Why two layers? Because your shell removes
the first layer as part of its argument parsing.

So to set a boolean:
tomato set is.this.sparta true citystates.toml

To set a string:
tomato set is.this.sparta '"true"' citystates.toml

Use the same approach for numbers.

Documented this feature in the readme.

This should fix # 2: Ability to set non-string values