feat: miette error messages, refactoring, and bugfixes
All the error messages now have helpful miette-driven suggestions
for how to fix them, if indeed they are fixable by the user. This
makes the code a lot busier, but the payoff is usability, so neat!
All error printing goes to stderr.

Took the opportunity to clean up some comments and some behavior
described in the comments. Removing keys is now idempotent: if you
try to remove a key that isn't present, this is not an error. Wrote
a test for this.

Refactored away repeated code in key traversal.